add play again button at the end.
add play again button and add it's functionality which is refresh the page,   is something i really don't like, idk what else to do for now but i will fixed it.

also change little bit of button style
and add a new div in html file for play again button to hold.