mainly update the CSS, with big different
in style.css a massive update to the game style, now it's not using a lot of VW and VH and % value, also style the new score div, change the avatar, health, health-bar, placeholder-dice, dice, and the new Play-Again button ,
also remove a lot of unnecessary style code and try to make it more simple and readable.

in index.js update the end-health class name to remining health,
in data.js add 10 (40) more health to ORC monster for better change to monster to win,
in Character class update the dice image alt with less text and more info,
in index.html it just format every things.